requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.3.0
multidict==6.0.5
mutagen==1.47.0
nltk==3.9.1
numpy==2.1.1
packaging==24.1
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
python-dotenv==1.0.1
redis==5.0.8
regex==2024.7.24
requests==2.32.3
spotipy==2.24.0
tqdm==4.66.5
urllib3==2.2.2
websockets==13.0.1
yarl==1.9.7
youtube-dl==2021.12.17
yt-dlp==2024.8.6
